fix(plugin-file-manager): move storage validation middleware
